massive progress on front end and backend

+++ server/Controllers/Section.js
@@ -1,5 +1,6 @@
 const Section = require("../Models/Section");
 const Course = require("../Models/Course");
+const SubSection = require("../Models/SubSection");
 
 // CREATE a new section
 exports.createSection = async (req, res) => {
@@ -15,7 +16,7 @@ exports.createSection = async (req, res) => {
             });
         }
 
-        // Create a new section with the given name  
+        // Create a new section with the given name
         const newSection = await Section.create({ sectionName });
 
         // Add the new section to the course's content array
@@ -52,50 +53,76 @@ exports.createSection = async (req, res) => {
     }
 };
 
+// UPDATE a section
 exports.updateSection = async (req, res) => {
     try {
-        //data input
-        const { sectionName, sectionId } = req.body;
-
-        //data validation
-        if (!sectionName || !sectionId) {
-            return res.status(400).json({
-                success: false,
-                message: "Missing Properties",
-            });
-        }
-
-        //update data
-        const section = await Section.findByIdAndUpdate(sectionId,
+        const { sectionName, sectionId, courseId } = req.body;
+        const section = await Section.findByIdAndUpdate(
+            sectionId,
             { sectionName },
-            { new: true },
+            { new: true }
         );
 
-        //return response
-        return res.status(200).json({
+        const course = await Course.findById(courseId)
+            .populate({
+                path: "courseContent",
+                populate: {
+                    path: "subSection",
+                },
+            })
+            .exec();
+
+        res.status(200).json({
             success: true,
-            message: "Section Updated successfully",
+            message: section,
+            data: course,
         });
-    }
-    catch (error) {
-        return res.status(500).json({
+    } catch (error) {
+        console.error("Error updating section:", error);
+        res.status(500).json({
             success: false,
-            message: "UNable to Update Section",
-            error: error.message,
+            message: "Internal server error",
         });
     }
-}
+};
 
 // DELETE a section
 exports.deleteSection = async (req, res) => {
     try {
-        //HW -> req.params -> test
-        const { sectionId } = req.params;
+
+        const { sectionId, courseId } = req.body;
+        await Course.findByIdAndUpdate(courseId, {
+            $pull: {
+                courseContent: sectionId,
+            }
+        })
+        const section = await Section.findById(sectionId);
+        console.log(sectionId, courseId);
+        if (!section) {
+            return res.status(404).json({
+                success: false,
+                message: "Section not Found",
+            })
+        }
+
+        //delete sub section
+        await SubSection.deleteMany({ _id: { $in: section.subSection } });
+
         await Section.findByIdAndDelete(sectionId);
-        //HW -> Course ko bhi update karo
+
+        //find the updated course and return 
+        const course = await Course.findById(courseId).populate({
+            path: "courseContent",
+            populate: {
+                path: "subSection"
+            }
+        })
+            .exec();
+
         res.status(200).json({
             success: true,
             message: "Section deleted",
+            data: course
         });
     } catch (error) {
         console.error("Error deleting section:", error);
@@ -104,4 +131,4 @@ exports.deleteSection = async (req, res) => {
             message: "Internal server error",
         });
     }
-};
+};
